+++ BetterHI3Launcher.csproj
@@ -122,6 +122,7 @@
     </ApplicationDefinition>
     <Compile Include="TextStrings_en.cs" />
     <Compile Include="TextStrings_ru.cs" />
+    <Compile Include="Utility.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
